// generated/pagopa/PagoPAUser.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { EmailAddress } from "./EmailAddress";
import * as t from "io-ts";
import { NonEmptyString } from "italia-ts-commons/lib/strings";

/**
 * User data needed by PagaPA proxy.
 */

// required attributes
const PagoPAUserR = t.interface({
  name: t.string,

  family_name: t.string,

  email: EmailAddress,

  mobile_phone: NonEmptyString
});

// optional attributes
const PagoPAUserO = t.partial({});

export const PagoPAUser = t.exact(
  t.intersection([PagoPAUserR, PagoPAUserO], "PagoPAUser")
);

export type PagoPAUser = t.TypeOf<typeof PagoPAUser>;